refactor(lock): align store interface and service behavior
