+++ deptrac.yaml
@@ -174,7 +174,6 @@ deptrac:
       - URI
     Cache:
       - I18n
-      - Lock
     Controller:
       - HTTP
       - Validation
